% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundScript.R
\name{bundScript}
\alias{bundScript}
\title{Include the Javascript Used in Shiny}
\usage{
bundScript(...)
}
\arguments{
\item{\dots}{There is no required arguments in this function. But user
could manipulate to include different css files.}
}
\value{
No return value. It is used to pass the Javascript to Shiny.
}
\description{
intermediate function used to include necessary javascript
to visualize tree structures and estimated treatment effect
in shiny
}
