% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalTree.branch.R
\name{causalTree.branch}
\alias{causalTree.branch}
\title{Compute the "branches" to be drawn for an \code{causalTree} object}
\usage{
causalTree.branch(x, y, node, branch)
}
\arguments{
\item{x}{covariates}

\item{y}{outcome}

\item{node}{node of the fitted tree}

\item{branch}{branch of the fitted tree}
}
\value{
number of branches to be drawn
}
\description{
Compute the "branches" to be drawn for an \code{causalTree} object
}
