% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalTree.matrix.R
\name{causalTree.matrix}
\alias{causalTree.matrix}
\title{Intermediate function for \code{causalTree}}
\usage{
causalTree.matrix(frame)
}
\arguments{
\item{frame}{inherited from data.frame}
}
\value{
A covariate matrix used in the causal regression.
}
\description{
Intermediate function for \code{causalTree}
}
