% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalTreecallback.R
\name{causalTreecallback}
\alias{causalTreecallback}
\title{Intermediate function for \code{causalTree}}
\usage{
causalTreecallback(mlist, nobs, init)
}
\arguments{
\item{mlist}{a list of user written methods}

\item{nobs}{number of observations}

\item{init}{function name}
}
\value{
split method written by users
}
\description{
This routine sets up the callback code for user-written split
routines in causalTree
}
