% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalTreeco.R
\name{causalTreeco}
\alias{causalTreeco}
\title{Intermediate function for \code{causalTree}}
\usage{
causalTreeco(tree, parms)
}
\arguments{
\item{tree}{an \code{causalTree} object}

\item{parms}{parms}
}
\value{
the x-y coordinates for a tree
}
\description{
Compute the x-y coordinates for a tree
}
