% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearTemp.R
\name{clearTemp}
\alias{clearTemp}
\title{Clear Temporary Files}
\usage{
clearTemp()
}
\value{
no return value, to unlink files under the temp folder
}
\description{
The files for shiny are saved in a temporary directory. The files can be
cleared manually using the `clearTemp()` function, or will automatically
be cleared when you close R
}
