% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.causalTree.R
\name{est.causalTree}
\alias{est.causalTree}
\title{Intermediate function for \code{causalTree}}
\usage{
est.causalTree(fit, x)
}
\arguments{
\item{fit}{an \code{causalTree} object}

\item{x}{covariates}
}
\value{
Intermediate estimation results for an \code{causalTree} object.
}
\description{
Run down the built tree and get the final leaf ids for estimation sample
}
