% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.causalTree.R
\name{estimate.causalTree}
\alias{estimate.causalTree}
\title{estimate causal Tree}
\usage{
estimate.causalTree(
  object,
  data,
  weights,
  treatment,
  na.action = na.causalTree
)
}
\arguments{
\item{object}{A tree-structured fit \code{rpart} object, such as one
generated as a \code{causalTree} fit.}

\item{data}{New data frame to be used for estimating effects within leaves.}

\item{weights}{optional case weights.}

\item{treatment}{The treatment status of observations in the new
dataframe, where 1 represents treated and 0 represents control.}

\item{na.action}{the default action deletes all observations for which
\code{y} is missing, but keeps those in which one or more predictors
are missing.}
}
\value{
Intermediate estimation results for an \code{causalTree} object
}
\description{
estimate causal Tree
}
\details{
When the leaf contains only treated or control cases, the function will
trace back to the leaf's parent node recursively until the parent can
be used to compute causal effect. Please see Athey and Imbens
\emph{Machine Learning Methods for Estimating Heterogeneous Causal
Effects} (2015) for details.
}
