% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatg.R
\name{formatg}
\alias{formatg}
\title{Intermediate function for \code{causalTree}}
\usage{
formatg(x, digits = getOption("digits"), format = paste0("\%.", digits, "g"))
}
\arguments{
\item{x}{input training data}

\item{digits}{number of digits to be kept}

\item{format}{format of exported vector}
}
\value{
No return value, called for formatting the exported estimates
}
\description{
Intermediate function for \code{causalTree}
}
