% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultPath.R
\name{getDefaultPath}
\alias{getDefaultPath}
\title{Get the Current Working Directory}
\usage{
getDefaultPath()
}
\value{
a temporary file path
}
\description{
get the current work directory and set it as the default directory to
save the shiny files temporarily
}
