% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDensities.R
\name{getDensities}
\alias{getDensities}
\title{Getting Distribution in Treatment and Control Groups}
\usage{
getDensities(treatment, outcome)
}
\arguments{
\item{treatment}{A character representing the name of treatment indicator.}

\item{outcome}{A character representing the name of outcome variable.}
}
\value{
vector of corresponding densities for each value of outcome vector
}
\description{
Getting the density of distribution in treatment and control groups,
which will be displayed in the
}
