% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honest.est.causalTree.R
\name{honest.est.causalTree}
\alias{honest.est.causalTree}
\title{honest re-estimation and change the frame of object using estimation sample}
\usage{
honest.est.causalTree(fit, x, wt, treatment, y)
}
\arguments{
\item{fit}{an \code{causalTree} object}

\item{x}{input training data}

\item{wt}{optional weights}

\item{treatment}{treatment variable}

\item{y}{outcome variable}
}
\value{
An object of class \code{rpart}.  See \code{\link{rpart.object}}.
}
\description{
honest re-estimation and change the frame of object using estimation sample
}
