% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honest.est.rparttree.R
\name{honest.est.rparttree}
\alias{honest.est.rparttree}
\title{honest re-estimation and change the frame of object using estimation sample}
\usage{
honest.est.rparttree(fit, x, wt, y)
}
\arguments{
\item{fit}{an \code{causalTree} object}

\item{x}{input training data}

\item{wt}{optional weights}

\item{y}{outcome variable}
}
\value{
Intermediate estimation results for an honest estimation of
\code{causalTree}.
}
\description{
honest re-estimation and change the frame of object using estimation sample
}
