% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte_plot.R
\name{hte_plot}
\alias{hte_plot}
\title{Visualize the Estimated Results}
\usage{
hte_plot(
  model,
  data,
  treatment_indicator = NULL,
  outcomevariable,
  propensity_score,
  plot.title = "Visualization of the Tree"
)
}
\arguments{
\item{model}{a tree model constructed by \code{hte_causalTree, hte_matchinleaves,
or hte_ipw}.}

\item{data}{a data frame containing the variables in the model.}

\item{treatment_indicator}{a character representing the column name
for the treatment variable in the causal setup.}

\item{outcomevariable}{a character representing the column name
of the outcome variable.}

\item{propensity_score}{a character representing the column name of
the propensity score.}

\item{plot.title}{character representing the main title of the plot.}
}
\value{
no return value
}
\description{
The  function \code{hte_plot} takes  a  model  created  by causal tree, as
well as the adjusted version, and
plots  the  distribution of the outcome variable in treated
and control groups in each leaf of the  tree.
This  visualization  aims  to  show  how  the  predicted
treatment  effect changes with each split in the tree.
}
