% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalTree.anova.R
\name{htetree.anova}
\alias{htetree.anova}
\title{Intermediate function for \code{causalTree}}
\usage{
htetree.anova(y, offset, wt)
}
\arguments{
\item{y}{outcome variable}

\item{offset}{this can be used to specify an a priori known
component
to be included in the linear predictor during fitting. This should be
\code{NULL} or a numeric vector of length equal to the number of cases.
One or more \link{offset} terms can be included in the formula instead or as
well, and
if more than one is specified their sum is used. See \link{model.offset}.}

\item{wt}{optional weights}
}
\value{
No return value.
}
\description{
Intermediate function for \code{causalTree}
}
