% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Caclulate variable importance}
\usage{
importance(fit)
}
\arguments{
\item{fit}{a fitted \code{causalTree} object.}
}
\value{
same as the \code{importance} function in \pkg{rpart}.
}
\description{
Each primary split is credited with the value of splits$improve
Each surrogate split gets split$adj times the primary split's value
}
