% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchinleaves.R
\name{matchinleaves}
\alias{matchinleaves}
\title{NN Matching in Leaves}
\usage{
matchinleaves(
  trainset = match_data,
  covariates = covariates,
  outcomevariable = outcomevariable,
  hte_effect_setup = hte_effect_setup,
  treatment_indicator,
  con.num = 1,
  ...
)
}
\arguments{
\item{trainset}{a data frame only containing the variables used in
the model and missings values are listwise deleted.}

\item{covariates}{a vector of column names of all
covariates (linear terms andpropensity score).}

\item{outcomevariable}{a character representing the column name
of the outcome variable.}

\item{hte_effect_setup}{a empty list to store the adjusted treatment
effect.}

\item{treatment_indicator}{a character representing the column
name of the treatment indicator.}

\item{con.num}{a number indicating the number of units from control
groups to be used in matching}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list for summarizing the results after matching.
}
\description{
This intermediate function is used to adjust the heterogeneous treatment
effect estimated in each leaf with NN matching.
}
