% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.frame.causalTree.R
\name{model.frame.causalTree}
\alias{model.frame.causalTree}
\title{Intermediate function for \code{causalTree}}
\usage{
\method{model.frame}{causalTree}(formula, ...)
}
\arguments{
\item{formula}{a \link{formula}, with a response but no interaction terms.
If this is a data frame, it is taken as the model frame (see
\link{model.frame}).}

\item{...}{arguments to \code{\link{rpart.control}} may also be
specified in the call to \code{causalTree}.  They are checked against the
list of valid arguments.  An example of a commonly set parameter would
be \code{xval}, which sets the number of cross-validation samples.
The parameter \code{minsize} is implemented differently in
\code{causalTree} than in {rpart}; we require a minimum of \code{minsize}
treated observations and a minimum of \code{minsize} control
observations in each leaf.}
}
\value{
a model frame for \code{causalTree}.
}
\description{
get model frame of causalTree, same as \link{rpart}
}
