% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.causalTree.R
\name{na.causalTree}
\alias{na.causalTree}
\title{Intermediate function for \code{causalTree}}
\usage{
na.causalTree(x)
}
\arguments{
\item{x}{covariates}
}
\value{
No return value, used for handling missing values when thy are
included in sample.
}
\description{
requirement when missing values are included in sample.
}
