% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDynamic.R
\name{runDynamic}
\alias{runDynamic}
\title{Visualize Causal Tree and Treatment Effects via Shiny}
\usage{
runDynamic(
  model,
  data,
  outcomevariable,
  treatment_indicator,
  propensity_score = ""
)
}
\arguments{
\item{model}{a tree model constructed by \code{hte_causalTree, hte_matchinleaves,
or hte_ipw}.}

\item{data}{a data frame containing the variables in the model.}

\item{outcomevariable}{a character representing the column name
of the outcome variable.}

\item{treatment_indicator}{a character representing the column name
for the treatment variable in the causal setup.}

\item{propensity_score}{a character representing the column name of
the propensity score.}
}
\value{
a Shiny page.
}
\description{
Visualize Causal Tree and Treatment Effects via Shiny
}
