% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveBCSS.R
\name{saveBCSS}
\alias{saveBCSS}
\title{Save Javascript Embedded in Shiny App}
\usage{
saveBCSS(filePath)
}
\arguments{
\item{filePath}{a character string representing the path name
to save the files temporarily.}
}
\value{
No return value. It is used to save necessary files temporarily
to run Shiny App.
}
\description{
Save Javascript Embedded in Shiny App
}
