% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGCSS.R
\name{saveGCSS}
\alias{saveGCSS}
\title{Save CSS File Embedded in Shiny App}
\usage{
saveGCSS(filePath)
}
\arguments{
\item{filePath}{a character string representing the path name
to save the files temporarily.}
}
\value{
No return value. It is used to save necessary files temporarily
to run Shiny App.
}
\description{
Save CSS File Embedded in Shiny App
}
