% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveInd.R
\name{saveInd}
\alias{saveInd}
\title{Save HTML Index Embedded in Shiny App}
\usage{
saveInd(filePath)
}
\arguments{
\item{filePath}{a character string representing the path name
to save the files temporarily.}
}
\value{
No return value. It is used to save necessary files temporarily
to run Shiny App.
}
\description{
Save HTML Index Embedded in Shiny App
}
