% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveServ.R
\name{saveServ}
\alias{saveServ}
\title{Save Shiny Server Temporarily}
\usage{
saveServ(filePath)
}
\arguments{
\item{filePath}{a character string representing the path name
to save the files temporarily.}
}
\value{
No return value. It is used to save necessary files temporarily
to run Shiny App.
}
\description{
Save Shiny Server Temporarily
}
