% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveUI.R
\name{saveUI}
\alias{saveUI}
\title{Save Shiny UI Temporarily}
\usage{
saveUI(filePath)
}
\arguments{
\item{filePath}{a character string representing the path name
to save the files temporarily.}
}
\value{
No return value. It is used to save necessary files temporarily
to run Shiny App.
}
\description{
Save Shiny UI Temporarily
}
