% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation.1}
\alias{simulation.1}
\title{A Simulated Dataset}
\format{
## `simulation.1`
A data frame with 500 observations on the following 12 variables.
\describe{
\item{\code{x1}}{a numeric vector}
\item{\code{x2}}{a numeric vector}
\item{\code{x3}}{a numeric vector}
\item{\code{x4}}{a numeric vector}
\item{\code{x5}}{a numeric vector}
\item{\code{x6}}{a numeric vector}
\item{\code{x7}}{a numeric vector}
\item{\code{x8}}{a numeric vector}
\item{\code{x9}}{a numeric vector}
\item{\code{x10}}{a numeric vector}
\item{\code{y}}{a numeric vector}
\item{\code{treatment}}{a numeric vector}
}
}
\usage{
simulation.1
}
\description{
A simulated dataset inherited from \code{causalTree} package
}
\keyword{datasets}
